+++ src/global/utils/progressbar.cpp
@@ -35,9 +35,9 @@ namespace pbar {
     auto it = std::find_if(units.begin(), units.end(), [&u](const auto& pr) {
       return pr.first == u;
     });
-    const int u_idx = (it != units.end()) ? std::distance(units.begin(), it) : -1;
+    const long u_idx = (it != units.end()) ? std::distance(units.begin(), it) : -1;
     raise::ErrorIf(u_idx < 0, "Invalid unit", HERE);
-    int shift = 0;
+    long shift = 0;
     if (t < 1) {
       shift = -1;
     } else if (1e3 <= t && t < 1e6) {
@@ -49,8 +49,8 @@ namespace pbar {
     } else if (3.6e9 <= t) {
       shift += 4;
     }
-    auto newu_idx = std::min(std::max(0, u_idx + shift),
-                             static_cast<int>(units.size()));
+    auto newu_idx = std::min(std::max(0L, u_idx + shift),
+                             static_cast<long>(units.size()));
     return { t * (units[u_idx].second / units[newu_idx].second),
              units[newu_idx].first };
   }
